+++ CodeConverter/CSharp/TypeConversionAnalyzer.cs
@@ -87,6 +87,9 @@ namespace ICSharpCode.CodeConverter.CSharp
         private ExpressionSyntax AddTypeConversion(VBSyntax.ExpressionSyntax vbNode, ExpressionSyntax csNode, TypeConversionKind conversionKind, bool addParenthesisIfNeeded, ITypeSymbol vbType, ITypeSymbol vbConvertedType)
         {
             switch (conversionKind) {
+                case TypeConversionKind.FractionalNumberRoundThenCast:
+                    csNode = AddRoundInvocation(vbNode, csNode, vbType, vbConvertedType);
+                    return AddTypeConversion(vbNode, csNode, TypeConversionKind.NonDestructiveCast, addParenthesisIfNeeded, vbType, vbConvertedType);
                 case TypeConversionKind.EnumConversionThenCast:
                     var underlyingType = ((INamedTypeSymbol) vbConvertedType).EnumUnderlyingType;
                     csNode = AddTypeConversion(vbNode, csNode, TypeConversionKind.Conversion, addParenthesisIfNeeded, vbType, underlyingType);
@@ -258,6 +261,9 @@ namespace ICSharpCode.CodeConverter.CSharp
 
             bool isConvertToString =
                         (vbConversion.IsString || vbConversion.IsReference && vbConversion.IsNarrowing)  && vbConvertedType.SpecialType == SpecialType.System_String;
+            bool isConvertFractionalToInt = 
+                        !csConversion.IsImplicit && vbType.IsFractionalNumericType() && vbConvertedType.IsNumericType() && !vbConvertedType.IsFractionalNumericType();
+
             if (!csConversion.Exists || csConversion.IsUnboxing) {
                 if (ConvertStringToCharLiteral(vbNode, vbType, vbConvertedType, out _)) {
                     typeConversionKind =
@@ -275,6 +281,9 @@ namespace ICSharpCode.CodeConverter.CSharp
                 }
             } else if (csConversion.IsExplicit && csConversion.IsEnumeration || csConversion.IsBoxing) {
                 typeConversionKind = TypeConversionKind.NonDestructiveCast;
+                return true;
+            } else if (vbConversion.IsNumeric && csConversion.IsNumeric && isConvertFractionalToInt) {
+                typeConversionKind = TypeConversionKind.FractionalNumberRoundThenCast;
                 return true;
             } else if (vbConversion.IsNumeric && csConversion.IsNumeric) {
                 // For widening, implicit, a cast is really only needed to help resolve the overload for the operator/method used.
@@ -308,7 +317,11 @@ namespace ICSharpCode.CodeConverter.CSharp
         {
             return _semanticModel.GetOperation(vbNode).Parent is IConversionOperation co && co.IsImplicit && co.Operand.ConstantValue.HasValue;
         }
-
+        private bool IsExplicitNumericConversion(VBSyntax.ExpressionSyntax vbNode)
+        {
+            return _semanticModel.GetOperation(vbNode).Parent is IConversionOperation co &&
+                   !co.IsImplicit; //&& co.Operand.i && co.Operand.ConstantValue.HasValue;
+        }
         private static TypeConversionKind AnalyzeVbConversion(bool alwaysExplicit, ITypeSymbol vbType,
             ITypeSymbol vbConvertedType, Conversion vbConversion)
         {
@@ -327,6 +340,17 @@ namespace ICSharpCode.CodeConverter.CSharp
             }
 
             return TypeConversionKind.Unknown;
+        }
+
+        public ExpressionSyntax AddRoundInvocation(VBSyntax.ExpressionSyntax vbNode, ExpressionSyntax csNode, ITypeSymbol currentType, ITypeSymbol targetType)
+        {
+ 
+            // Add to specify we're using System for Math.Round
+            _extraUsingDirectives.Add("System");
+            var memberAccess = SyntaxFactory.MemberAccessExpression(SyntaxKind.SimpleMemberAccessExpression,
+                SyntaxFactory.IdentifierName("Math"),  SyntaxFactory.IdentifierName("Round"));
+            var arguments = SyntaxFactory.ArgumentList(SyntaxFactory.SingletonSeparatedList(SyntaxFactory.Argument(csNode)));
+            return SyntaxFactory.InvocationExpression(memberAccess, arguments);
         }
 
         public ExpressionSyntax AddExplicitConvertTo(VBSyntax.ExpressionSyntax vbNode, ExpressionSyntax csNode, ITypeSymbol currentType, ITypeSymbol targetType)
@@ -385,7 +409,8 @@ namespace ICSharpCode.CodeConverter.CSharp
             EnumCastThenConversion,
             NullableBool,
             StringToCharArray,
-            DelegateConstructor
+            DelegateConstructor,
+            FractionalNumberRoundThenCast
         }
 
         public static bool ConvertStringToCharLiteral(VBSyntax.ExpressionSyntax node,