+++ Tests/CSharp/ExpressionTests/ExpressionTests.cs
@@ -148,7 +148,7 @@ public partial class EnumToString
     private void TEest2(Tes aEnum)
     {
         string sxtr_Tmp = ""Use"" + ((short)aEnum).ToString();
-        short si_Txt = (short)Math.Pow(2d, (double)Tes.TEST2);
+        short si_Txt = (short)Math.Round(Math.Pow(2d, (double)Tes.TEST2));
     }
 }");
         }
@@ -1581,18 +1581,19 @@ public partial class MoreParsing
         anInt += aDec
     End Sub
 End Class",
-                @"
+                @"using System;
+
 public partial class Compound
 {
     public void Operators()
     {
         int anInt = 123;
         decimal aDec = 12.3m;
-        anInt = (int)(anInt * aDec);
-        anInt = (int)(anInt / (long)aDec);
-        anInt = (int)(anInt / aDec);
-        anInt = (int)(anInt - aDec);
-        anInt = (int)(anInt + aDec);
+        anInt = (int)Math.Round(anInt * aDec);
+        anInt = (int)(anInt / (long)Math.Round(aDec));
+        anInt = (int)Math.Round(anInt / aDec);
+        anInt = (int)Math.Round(anInt - aDec);
+        anInt = (int)Math.Round(anInt + aDec);
     }
 }");
         }
@@ -1612,18 +1613,19 @@ public partial class Compound
         aShort += aDec
     End Sub
 End Class",
-                @"
+                @"using System;
+
 public partial class Compound
 {
     public void Operators()
     {
         short aShort = 123;
         decimal aDec = 12.3m;
-        aShort = (short)(aShort * aDec);
-        aShort = (short)(aShort / (long)aDec);
-        aShort = (short)(aShort / aDec);
-        aShort = (short)(aShort - aDec);
-        aShort = (short)(aShort + aDec);
+        aShort = (short)Math.Round(aShort * aDec);
+        aShort = (short)(aShort / (long)Math.Round(aDec));
+        aShort = (short)Math.Round(aShort / aDec);
+        aShort = (short)Math.Round(aShort - aDec);
+        aShort = (short)Math.Round(aShort + aDec);
     }
 }");
         }
@@ -1643,7 +1645,8 @@ public partial class Compound
         aShort += anInt
     End Sub
 End Class",
-                @"
+                @"using System;
+
 public partial class Compound
 {
     public void Operators()
@@ -1652,7 +1655,7 @@ public partial class Compound
         int anInt = 12;
         aShort = (short)(aShort * anInt);
         aShort = (short)(aShort / anInt);
-        aShort = (short)(aShort / (double)anInt);
+        aShort = (short)Math.Round(aShort / (double)anInt);
         aShort = (short)(aShort - anInt);
         aShort = (short)(aShort + anInt);
     }
@@ -1687,6 +1690,28 @@ public partial class Compound
         }
 
         [Fact]
+        public async Task CintIsConvertedCorrectly()
+        {
+            await TestConversionVisualBasicToCSharpAsync(
+                @"Public Class Compound
+    Public Sub Operators()
+        Dim do_Tmp As Double = 9999 / 100
+        Dim i_Tmp as Integer = CInt(do_Tmp)
+    End Sub
+End Class",
+                @"using System;
+
+public partial class Compound
+{
+    public void Operators()
+    {
+        double do_Tmp = 9999d / 100d;
+        int i_Tmp = (int)Math.Round(do_Tmp);
+    }
+}");
+        }
+
+        [Fact]
         public async Task ArgumentsAreTypeConvertedAsync()
         {
             await TestConversionVisualBasicToCSharpAsync(
@@ -1698,14 +1723,15 @@ Public Class Compound
         Dim arry = New Single(7/someInt) {}
     End Sub
 End Class",
-                @"using System.Drawing;
+                @"using System;
+using System.Drawing;
 
 public partial class Compound
 {
     public void TypeCast(int someInt)
     {
-        var col = Color.FromArgb((int)(someInt * 255.0f), (int)(someInt * 255.0f), (int)(someInt * 255.0f));
-        var arry = new float[(int)(7d / someInt + 1)];
+        var col = Color.FromArgb((int)Math.Round(someInt * 255.0f), (int)Math.Round(someInt * 255.0f), (int)Math.Round(someInt * 255.0f));
+        var arry = new float[(int)Math.Round(7d / someInt + 1)];
     }
 }");
         }